switch from in to float

+++ libs/PhStrip/PhStripPeopleObject.cpp
@@ -7,7 +7,7 @@
 #include "PhStrip/PhStripPeopleObject.h"
 
 
-PhStripPeopleObject::PhStripPeopleObject(PhTime timeIn, PhPeople *people, PhTime timeOut, int track) : PhStripObject(timeIn),
+PhStripPeopleObject::PhStripPeopleObject(PhTime timeIn, PhPeople *people, PhTime timeOut, float track) : PhStripObject(timeIn),
 	_timeOut(timeOut), _people(people), _track(track)
 {
 }
